+++ Changelog.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+<<<<<<< Updated upstream
+=======
+## [0.0.24 - Beta] - 2022-12-20
+
+# Added
+- themeability
+Themes can be added by copying the default.theme file in the themes subfolder or /themes for docker to xxx.theme and editing the variables then changing the theme in config.conf
+
 ## [0.0.23 - Beta] - 2022-12-18
 
 # Fixed
@@ -14,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 # Added
 - Speed improvements
 
+>>>>>>> Stashed changes
 ## [0.0.21 - Beta] - 2022-11-20
 
 # Added